+++ src/main/java/com/nhom7/coworkingspace/service/BookingService.java
@@ -6,7 +6,6 @@ import com.nhom7.coworkingspace.dto.request.BookingSearchRequest;
 import com.nhom7.coworkingspace.dto.response.BookingResponse;
 import com.nhom7.coworkingspace.dto.response.PageResponse;
 import com.nhom7.coworkingspace.entity.Booking;
-import com.nhom7.coworkingspace.enums.BookingStatus;
 
 public interface BookingService {
 
@@ -22,7 +21,7 @@ public interface BookingService {
 
     BookingResponse getBookingById(Long bookingId);
 
-    BookingResponse updateBookingStatusByHost(Long bookingId, BookingStatus newStatus, String hostEmail);
+    BookingResponse updateBookingStatusByHost(Long bookingId, String newStatus, String hostEmail);
 
     PageResponse<BookingResponse> getBookingsForHost(String hostEmail, int page, int size);
 }